Rename GoldShop -> CurrencyShop, more modular
Supports Giving the player Gold for Silver, or Gold for "RealMoney" (nothing)

Refactors bits of the code to be more DRY-compliant
